Require syslog only when necessary
Requiring syslog on Ruby 3.3.0 results in an warning if the gem is not
explicitly installed. Suppress it by requiring syslog only when the user
tries to use Logging::Appender::Syslog.
